implement remote ai worker: implement O+T communication, move processing to Transcoder
